Refactor Predicate classes into predicates folder
