+++ src/main/java/seedu/address/model/person/predicates/NricMatchesPredicate.java
@@ -1,8 +1,9 @@
-package seedu.address.model.person;
+package seedu.address.model.person.predicates;
 
 import java.util.function.Predicate;
 
 import seedu.address.commons.util.ToStringBuilder;
+import seedu.address.model.person.Person;
 
 /**
  * Tests that a {@code Person}'s {@code NRIC} matches the given NRIC exactly.
