+++ src/main/java/seedu/address/model/person/predicates/TagContainsKeywordsPredicate.java
@@ -1,7 +1,9 @@
-package seedu.address.model.person;
+package seedu.address.model.person.predicates;
 
 import java.util.List;
 import java.util.function.Predicate;
+
+import seedu.address.model.person.Person;
 
 /**
  * Tests that a {@code Person}'s {@code Tag} matches any of the keywords given.
