+++ src/test/java/seedu/address/logic/commands/ViewCommandTest.java
@@ -12,7 +12,7 @@ import org.junit.jupiter.api.Test;
 import seedu.address.model.Model;
 import seedu.address.model.ModelManager;
 import seedu.address.model.UserPrefs;
-import seedu.address.model.person.NricMatchesPredicate;
+import seedu.address.model.person.predicates.NricMatchesPredicate;
 
 /**
  * Contains integration tests (interaction with the Model) for {@code ViewCommand}.
