+++ src/test/java/seedu/address/model/person/NricMatchesPredicateTest.java
@@ -5,6 +5,7 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import org.junit.jupiter.api.Test;
 
+import seedu.address.model.person.predicates.NricMatchesPredicate;
 import seedu.address.testutil.PersonBuilder;
 
 public class NricMatchesPredicateTest {
